We do not want to require any particular Julia package when we load JuliaExperimental. (Of course most code of the package will not be available if no Julia packages are available.) For that, `gap/utils.gi` and `julia/utils.jl` have been split into a Nemo independent part and a part which requires Nemo. (This addresses issue #65.)
The same holds for the test files,
we cannot expect that all tests can be run with a given installation
of Julia.
Therefore, `TestDirectory` has to be run just on those test files
for which the necessary Julia packages are available.

Besides the general test mechanism,
also some tests have now been adjusted to Julia 1.0.
