+++ JuliaExperimental/tst/utils.tst
@@ -1,24 +1,12 @@
 #############################################################################
 ##
-#W  utils.tst         GAP 4 package JuliaExperimental          Thomas Breuer
+#W  utils.tst         GAP 4 package JuliaExperimental           Thomas Breuer
 ##
 gap> START_TEST( "utils.tst" );
 
-##  small or large integers
-gap> l:= JuliaArrayOfFmpz( [ -2, -1, 0, 1, 2, 3 ] );
-<Julia: Nemo.fmpz[-2, -1, 0, 1, 2, 3]>
-gap> l:= JuliaArrayOfFmpz( [ 1, 2, 3, 2^70 ] );
-<Julia: Nemo.fmpz[1, 2, 3, 1180591620717411303424]>
-
-##  small or large rationals
-gap> l:= JuliaArrayOfFmpq( [ -2, -1/2, 0, 1, 2/3, 3/7 ] );
-<Julia: Nemo.fmpq[-2, -1//2, 0, 1, 2//3, 3//7]>
-gap> l:= JuliaArrayOfFmpq( [ 2^70/3, 1/2^70 ] );
-<Julia: Nemo.fmpq[1180591620717411303424//3, 1//1180591620717411303424]>
-
 ##
-gap> RecNames( CallFuncListWithTimings( Julia.Base.sleep, [ 2 ] ) );
-[ "result", "GAP_time", "Julia_time" ]
+gap> JuliaMatrixFromGapMatrix( [ [ 1, 2 ], [ 3, 4 ] ] );
+<Julia: [1 2; 3 4]>
 
 ##
 gap> STOP_TEST( "utils.tst", 1 );
